video: tegra: nvmap: Correct the data type
When granule_size of type u32 and size is more than 4GB (i.e. value
having more than 32 bits), then ALIGN_GRANULE_SIZE would return
incorrect value due to ~(GRANULE_SIZE - 1),  as this would be in 32
bits. Hence change the granule size's data type to size_t, so that the
above value is not restricted to 32 bits.

Bug 4650102

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/3149162
Tested-by: Bitan Biswas <bbiswas@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>

+++ drivers/video/tegra/nvmap/nvmap_init.c
@@ -498,7 +498,7 @@ void nvmap_dma_free_attrs(struct device *dev, size_t size, void *cpu_addr,
 	struct dma_coherent_mem_replica *mem;
 	bool is_gpu = false;
 	const char *device_name;
-	u32 granule_size = 0;
+	size_t granule_size = 0;
 
 	if (!dev || !dev->dma_mem)
 		return;
